requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
torch
torchvision
pinecone
google-generativeai
pillow
numpy
pandas
python-dotenv
aiofiles
requests